+++ lib/element_params/sub/button/checkbox_params.dart
@@ -33,5 +33,5 @@ class CheckboxParams {
   SelectableIconBtnParams? enableParams;
   SelectableIconBtnParams? disableParams;
   List<bool>? checkValues;
-  List<void Function(bool isEnabled)?>? callbacks;
+  void Function(List<bool>? checkValues)? callback;
 }